requirements.txt
# MarketFinder ETL - Python Requirements
# Generated from pyproject.toml for environments that don't support Poetry

# Core dependencies
polars>=0.20.0
pydantic>=2.5.0
duckdb>=0.10.0

# Async HTTP
httpx>=0.26.0
aiohttp>=3.9.0

# Orchestration
apache-airflow>=2.8.0

# ML and Analytics
scikit-learn>=1.4.0
numpy>=1.26.0

# LLM Integration
openai>=1.10.0
google-cloud-aiplatform>=1.40.0

# Caching and Storage
redis>=5.0.0

# Configuration and Environment
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Logging and Monitoring
structlog>=23.2.0
prometheus-client>=0.19.0

# Utilities
rich>=13.7.0
typer>=0.9.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=23.12.0
ruff>=0.1.9
mypy>=1.8.0
ipython>=8.18.0
jupyter>=1.0.0
pre-commit>=3.6.0